requirements.txt
-f https://download.pytorch.org/whl/cu124
torch
torchvision
numpy~=2.2.2
tifffile~=2025.1.10
matplotlib~=3.10.0
opencv-python~=4.11.0.86
pillow~=11.0.0
torch
torchvision
scipy~=1.15.1